% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{as.numeric,sparse_numeric-method}
\alias{as.numeric,sparse_numeric-method}
\title{Coerce sparse_numeric to numeric via as.numeric()}
\usage{
\S4method{as.numeric}{sparse_numeric}(x)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}
}
\value{
A numeric vector of length \code{x@length}.
}
\description{
Coerce sparse_numeric to numeric via as.numeric()
}
