% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{norm}
\alias{norm}
\alias{norm,sparse_numeric-method}
\title{Euclidean norm of a sparse_numeric vector}
\usage{
norm(x, ...)

\S4method{norm}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{...}{Ignored; included for compatibility with the generic.}
}
\value{
Numeric scalar norm.
}
\description{
Computes the Euclidean (L2) norm
\deqn{\sqrt{\sum_i x_i^2}}.
}
