% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{Dot product (cross product) of two sparse_numeric vectors}
\usage{
sparse_crossprod(x, y, ...)

\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x, y}{\code{sparse_numeric} objects.}

\item{...}{Ignored; included for method compatibility.}
}
\value{
A numeric scalar equal to the dot product.
}
\description{
Computes the inner product \eqn{\sum_i x_i y_i}.
}
