% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_mult}
\alias{sparse_mult}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\title{Elementwise product of two sparse_numeric vectors}
\usage{
sparse_mult(x, y, ...)

\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x, y}{\code{sparse_numeric} objects.}

\item{...}{Ignored; included for method compatibility.}
}
\value{
A \code{sparse_numeric} representing the elementwise product.
}
\description{
Multiplies corresponding elements of two sparse vectors.
}
