% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{standardize}
\alias{standardize}
\alias{standardize,sparse_numeric-method}
\title{Standardize a sparse_numeric vector}
\usage{
standardize(x, ...)

\S4method{standardize}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{...}{Ignored; included for compatibility with the generic.}
}
\value{
A \code{sparse_numeric} object containing standardized values.
}
\description{
Standardize a sparse_numeric vector
}
